Tracked down and corrected a nasty bug in which the Sass compiler would choke on the informational CSS comments being output in the size-class() mixin (e.g. "Conditional Layout: (columns: 640, gutters: 0)")—see changes to _size-class.scss in this commit.
The root cause: I was using Sass's interpolation feature (http://sass-lang.com/documentation/file.SASS_REFERENCE.html#interpolation_) to output the value of two Sass variables in the comments. This interpolation feature is only for use with CSS values. Hence the error we were seeing when compiling:

    Error: (columns: 640, gutters: 0) isn't a valid CSS value.\n        on line 135 of node_modules/susy/sass/susy/language/susy/_settings.scss

This isn't the right way to do this; replacing the #{} usage with a call to Sass's inspect() function is the correct way to output the literal content of these variables into the comment. More on Sass's inspect() function here: http://sass-lang.com/documentation/Sass/Script/Functions.html#inspect-instance_method