Reviewed fix for #11 and resolved bad side effects
Way back when, "Fix #11" in commit 7670f30 resolved the issue in two ways:

1. Changed the order of operations when determining the minimum font size for a call to `limit-font-size()` when outputting an `auto` minimum font size

2. Introduced `@at-root (without: media)`, with the intention of ensuring that the media query for the final font size limit would be set without any surrounding media queries preventing the style from applying

(1)—the order of operations change—appears to resolve most of the problems that Issue #11 documented.

(2), on the other hand, seems to have been a horrible design error, in the sense that a developer uses media queries, usually by way of the `size-class()` mixin, to constrain the range of widths that the `font-size()` mixin call applies to. Using `@at-root (without: media)` resulted in unintentional and hard-to-predict mixin behavior that was hard to work around and hard to even describe.

Therefore, this update removes (2) (the `@at-root (without: media)` wrap). As a result, any min and max font sizes that would be rendered by using `font-size()` in `vw` output mode should now be correctly contained by any surrounding media query.

This is by far more intuitive behavior from a developer perspective and will make this mixin easier to use and reason about.