Fix width values returned when pspx-ratio is 1
Several utility functions always assumed the size-class config file's pspx ratio was equal to the "actual" ratio (e.g. config-pspx-ratio = pspx-ratio = 2). This is not always the case of course, so for some values there needs to be pre or post ratio multiplication.

The fixes in this commit have been informally tested with pspx-ratio values of 2 (the default) and 1. These functions/mixins were fixed:

- Added _nytpi-sass-utils-number-functions-get-config-pixel-ratio() function
- _nytpi-sass-utils-vw() function: the $unitless input value is now scaled by the ratio pspx-ratio / config-pspx-ratio
- get-pspx-width() function: when returning a column width from $size-class-config, the number of columns for the corresponding pspx layout is now scaled by the ratio pspx-ratio / config-pspx-ratio (this scale factor is cancelled out if get-pspx-width is outputting in vw units)